Ties services with themselves as dependencies
When the node joins to the master device, the script controls the
order in which services are started.
When starting the system, the same order may be different each time.
This causes the start of the services may fail. For example, the kubelet cannot
be started when the docker is not running.

To control the order in which services start during startup, the
services are further configured as dependencies.