Scripts update to make them usable with hostprocess model
It seems that Install-Container.ps1 and PrepareNode.ps1 may be useful to
build Windows node with HostProcess containers and k8s 1.25.3 at least.
But some changes are needed, becasue some code does not cooperate.
This commit is a propose to leave current code unchanged for backward compatibility
and introduce UseHostProcess parameter to handle code not needed
to support HostProcess containers where are engaged.

Additionally it offers SuppressHints parameter to have an option
to omit on demand some help messages. Omitted code depends on git tool installed.
It may be an optimization for automated node provisioning (eg. vagrant)
and saves some time wasted for git install and repository cloning.