Updating Windows Scripts (#262)
* Fix guide errors

Fixing flannel rbac file location and removing field not necesary.
Also, updating Calico.yml to have calico-version everywhere!

* Updating Windows Scripts

Update for Windows scripts InstallContainerd.ps1 and PrepareNode.ps1.
Changed the location for containerd to the correct one in the script and changed the script which installs kubeadm, kubelet to k8s version v1.25.3 and removed Docker references since it is deprecated!