docs: Consistent format of versions input for scripts
This make versions input a no-brainer as currently it may alert
users with unnecessary questions "is that a typo? which one is correct? shall I use `v` prefix for the ContainerD installer?", etc.

The `PrepareNode.ps1` can handle both formats `X.Y.Z` and `vX.Y.Z`
though, but one has to learn it.